rag_flask_app_backend/requirements.txt
langchain
langchain_groq
langchain-community
python-dotenv
flask
flask_cors
psycopg2-binary
openai
tiktoken
faiss-gpu
gunicorn